rating sort and some modularization
